Fix finding rev in hg repo with no outgoing commits
The script seemed to be all right at finding a published ancestor if the
current workdir was an outgoing commit. However, in case the checked-out
commit was already published, the `hg log` command came up empty.
`output.split()[:2]` would then throw an error because `output` was the
empty string.

This patch alleviates that by including the checked-out commit in the
list of outgoing commits to find a common ancestor for.