Improve gecko gdb support
Without these changes, when run against a mozilla-central (gecko)
source tree, no gdb files will be found.

With these changes, the gdbs will be found and an attempt made to
load them, but the `import gdbpp` line will fail on
`gdb.lookup_type("uintptr_t")` because the .gdbinit will be loaded
before such a symbol is known.

Gecko users can work-around this issue by typing `python import gdbpp`
later on and then everything will be fine.  Or that problem could be
fixed, but arguably this is an improvement.

Because it seems like ".gdbinit" might have inadvertantly become load
bearing even though the comment said it was for "mozilla", I've left
it in.